chore(frontend): inline pure-wrapper Page components
Each *Page.tsx that was just <PageShell><Inner/></PageShell> now contains
the inner body directly. For param-resolving pages (Members, JoinRequests,
InviteMember), the route param and group lookup moved into the Page
component, dropping the props interface on the inner.

Affects: Blocked, Preferences, Requests, SearchGroup, Settings, Invites,
AllJoinRequests, JoinRequests, Members, InviteMember.
