+++ frontend/src/pages/SettingsPage.tsx
@@ -1,11 +1,594 @@
-import React from "react";
+import React, { useState, useEffect, useCallback, useRef } from "react";
+import { Upload, User } from "lucide-react";
+import { useQueryClient } from "@tanstack/react-query";
 import { PageShell } from "../components/Layout/PageShell";
-import { Settings } from "./Settings";
+import { useAppStore } from "../stores/appStore";
+import { uploadAvatar } from "../services/r2-upload";
+import { resizeImage } from "../utils/imageProcessing";
+import { useUserProfile, useUpdateProfile, useUpdateAvatar, useUserAvatar, userQueryKeys } from "../hooks/queries";
+import { TextInput } from "../components/ui/TextInput";
+import { Button } from "../components/ui/Button";
+import { getVersion } from "@tauri-apps/api/app";
+import { check as checkForUpdate } from "@tauri-apps/plugin-updater";
+import { invoke } from "@tauri-apps/api/core";
 
 export const SettingsPage: React.FC = () => {
+  const { currentUser } = useAppStore();
+  const setUpdateRequired = useAppStore((s) => s.setUpdateRequired);
+
+  const { data: userData, isLoading } = useUserProfile();
+  const { data: avatarDownloadUrl } = useUserAvatar();
+  const updateProfileMutation = useUpdateProfile();
+  const updateAvatarMutation = useUpdateAvatar();
+
+  const [selectedFile, setSelectedFile] = useState<File | null>(null);
+  const [preview, setPreview] = useState<string | null>(null);
+  const [currentAvatarUrl, setCurrentAvatarUrl] = useState<string | null>(null);
+  const [uploadError, setUploadError] = useState<string | null>(null);
+  const [username, setUsername] = useState("");
+  const [preferredName, setPreferredName] = useState("");
+  const [email, setEmail] = useState("");
+  const [phone, setPhone] = useState("");
+
+  // Email-change flow lives entirely in this component:
+  //   'idle'    → display current email + "Change" button
+  //   'request' → input new address, send OTP
+  //   'verify'  → input the OTP, swap email
+  // Distinct mutations would obscure the linear UX, so we hold state locally.
+  const [emailChangeStep, setEmailChangeStep] = useState<"idle" | "request" | "verify">("idle");
+  const [pendingNewEmail, setPendingNewEmail] = useState("");
+  const [emailOtpCode, setEmailOtpCode] = useState("");
+  const [emailChangeError, setEmailChangeError] = useState<string | null>(null);
+  const [emailChangePending, setEmailChangePending] = useState(false);
+  const queryClient = useQueryClient();
+  const [fileInputKey, setFileInputKey] = useState(0);
+  const fileInputRef = useRef<HTMLInputElement>(null);
+  const [saveSuccess, setSaveSuccess] = useState(false);
+  const [appVersion, setAppVersion] = useState<string>("");
+  const [updateStatus, setUpdateStatus] = useState<"idle" | "checking" | "available" | "none" | "error">("idle");
+  const [updateVersion, setUpdateVersion] = useState<string>("");
+  const [updateError, setUpdateError] = useState<string>("");
+  const [isInstalling, setIsInstalling] = useState(false);
+
+  useEffect(() => {
+    return () => { if (preview) { URL.revokeObjectURL(preview); } };
+  }, [preview]);
+
+  useEffect(() => {
+    if (userData) {
+      setUsername(userData.username || "");
+      setPreferredName(userData.preferred_name || "");
+      setEmail(userData.email || "");
+      setPhone(userData.phone || "");
+    }
+  }, [userData]);
+
+  useEffect(() => {
+    getVersion().then(setAppVersion).catch(() => setAppVersion("unknown"));
+  }, []);
+
+  useEffect(() => { setCurrentAvatarUrl(avatarDownloadUrl || null); }, [avatarDownloadUrl]);
+
+  // Accept image files dropped anywhere on the window while on this page.
+  useEffect(() => {
+    const handlePathDrop = (e: Event) => {
+      const paths: string[] = (e as CustomEvent<{ paths: string[] }>).detail?.paths ?? [];
+      const imagePath = paths.find((p) => /\.(png|jpe?g|gif|webp|avif|svg)$/i.test(p));
+      if (!imagePath) {
+        return;
+      }
+      // Convert the native path to a File-like object via fetch(convertFileSrc(path)).
+      // Tauri exposes native paths through the asset protocol.
+      import("@tauri-apps/api/core").then(({ convertFileSrc }) => {
+        const src = convertFileSrc(imagePath);
+        fetch(src)
+          .then((r) => r.blob())
+          .then((blob) => {
+            const name = imagePath.split(/[\\/]/).pop() ?? "image";
+            const file = new File([blob], name, { type: blob.type || "image/png" });
+            setSelectedFile(file);
+            setUploadError(null);
+            if (preview) {
+              URL.revokeObjectURL(preview);
+            }
+            setPreview(URL.createObjectURL(file));
+          })
+          .catch((err) => {
+            console.error("[Settings] pathdrop read failed:", err);
+          });
+      });
+    };
+
+    window.addEventListener("pollis:pathdrop", handlePathDrop);
+    return () => window.removeEventListener("pollis:pathdrop", handlePathDrop);
+  }, [preview]);
+
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+    if (!file) {
+      setSelectedFile(null);
+      if (preview) { URL.revokeObjectURL(preview); }
+      setPreview(null);
+      setUploadError(null);
+      return;
+    }
+    setSelectedFile(file);
+    setUploadError(null);
+    if (preview) { URL.revokeObjectURL(preview); }
+    if (file.type.startsWith("image/")) { setPreview(URL.createObjectURL(file)); }
+  };
+
+  const handleAvatarUpload = useCallback(async () => {
+    if (!selectedFile || !currentUser) { return; }
+    setUploadError(null);
+    try {
+      const optimizedFile = await resizeImage(selectedFile);
+      const response = await uploadAvatar(currentUser.id, "", optimizedFile);
+      await updateAvatarMutation.mutateAsync(response.object_key);
+      setSelectedFile(null);
+      if (preview) { URL.revokeObjectURL(preview); }
+      setPreview(null);
+      setFileInputKey((prev) => prev + 1);
+      setSaveSuccess(true);
+    } catch (error) {
+      setUploadError(error instanceof Error ? error.message : "Failed to upload avatar");
+    }
+  }, [selectedFile, currentUser, preview, updateAvatarMutation]);
+
+  useEffect(() => {
+    if (saveSuccess) {
+      const t = setTimeout(() => setSaveSuccess(false), 3000);
+      return () => clearTimeout(t);
+    }
+  }, [saveSuccess]);
+
+  const handleSave = async () => {
+    if (!currentUser) { return; }
+    try {
+      await updateProfileMutation.mutateAsync({
+        username: username.trim(),
+        preferredName: preferredName.trim() || undefined,
+        phone: phone.trim() || undefined,
+      });
+      setSaveSuccess(true);
+      setTimeout(() => setSaveSuccess(false), 3000);
+    } catch (error) {
+      console.error("Failed to save settings:", error);
+    }
+  };
+
+  const cancelEmailChange = () => {
+    setEmailChangeStep("idle");
+    setPendingNewEmail("");
+    setEmailOtpCode("");
+    setEmailChangeError(null);
+  };
+
+  const handleSendEmailChangeOtp = async () => {
+    if (!currentUser) { return; }
+    const target = pendingNewEmail.trim();
+    if (!target) {
+      setEmailChangeError("Enter a new email address.");
+      return;
+    }
+    setEmailChangePending(true);
+    setEmailChangeError(null);
+    try {
+      await invoke("request_email_change_otp", { userId: currentUser.id, newEmail: target });
+      setEmailChangeStep("verify");
+    } catch (err) {
+      setEmailChangeError(err instanceof Error ? err.message : String(err));
+    } finally {
+      setEmailChangePending(false);
+    }
+  };
+
+  const handleVerifyEmailChange = async () => {
+    if (!currentUser) { return; }
+    if (!emailOtpCode.trim()) {
+      setEmailChangeError("Enter the code from your email.");
+      return;
+    }
+    setEmailChangePending(true);
+    setEmailChangeError(null);
+    try {
+      await invoke("verify_email_change", {
+        userId: currentUser.id,
+        newEmail: pendingNewEmail.trim(),
+        code: emailOtpCode.trim(),
+      });
+      // Refetch the profile so the displayed email updates without a reload.
+      await queryClient.invalidateQueries({ queryKey: userQueryKeys.profile(currentUser.id) });
+      cancelEmailChange();
+    } catch (err) {
+      setEmailChangeError(err instanceof Error ? err.message : String(err));
+    } finally {
+      setEmailChangePending(false);
+    }
+  };
+
+  const handleCheckForUpdates = useCallback(async () => {
+    setUpdateStatus("checking");
+    setUpdateError("");
+    setUpdateVersion("");
+    try {
+      const update = await checkForUpdate();
+      if (update) {
+        setUpdateStatus("available");
+        setUpdateVersion(update.version);
+      } else {
+        setUpdateStatus("none");
+      }
+    } catch (error) {
+      setUpdateStatus("error");
+      setUpdateError(error instanceof Error ? error.message : "Failed to check for updates");
+    }
+  }, []);
+
+  const handleInstallUpdate = useCallback(async () => {
+    if (updateStatus !== "available") {
+      return;
+    }
+    setIsInstalling(true);
+    // Mark the Rust-side flag and flip the store so App.tsx takes over
+    // with the fullscreen UpdateScreen. The screen handles graceful
+    // voice/network teardown, download, install, and relaunch.
+    await invoke("mark_update_required").catch(() => {});
+    setUpdateRequired(true);
+  }, [updateStatus, setUpdateRequired]);
+
+  if (!currentUser) {
+    return (
+      <PageShell title="User Settings" scrollable>
+        <div data-testid="settings-no-user" className="flex items-center justify-center flex-1" style={{ background: 'var(--c-bg)' }}>
+          <p className="text-xs font-mono" style={{ color: 'var(--c-text-muted)' }}>Please sign in</p>
+        </div>
+      </PageShell>
+    );
+  }
+
   return (
     <PageShell title="User Settings" scrollable>
-      <Settings />
+    <div
+      data-testid="settings-page"
+      className="flex-1 flex flex-col overflow-auto"
+      style={{ background: 'var(--c-bg)' }}
+    >
+      <div data-testid="settings-content" className="flex-1 flex justify-center overflow-auto px-6 py-8">
+        <div className="w-full max-w-md flex flex-col gap-8">
+
+          {/* Account */}
+          <section className="flex flex-col gap-4 mb-12">
+            <h2 className="text-xs font-mono font-medium uppercase tracking-widest pb-1 border-b" style={{ color: 'var(--c-text-dim)', borderColor: 'var(--c-border)' }}>
+              Account
+            </h2>
+
+            {isLoading ? (
+              <span data-testid="settings-loading" className="text-xs font-mono" style={{ color: 'var(--c-text-muted)' }}>
+                Loading…
+              </span>
+            ) : (
+              <div className="flex flex-col gap-4">
+                <TextInput
+                  label="Username"
+                  value={username}
+                  onChange={setUsername}
+                  placeholder="username"
+                  id="settings-username"
+                />
+                <input data-testid="settings-username-input" type="hidden" value={username} readOnly />
+
+                <TextInput
+                  label="Preferred name"
+                  value={preferredName}
+                  onChange={setPreferredName}
+                  placeholder="What people call you"
+                  id="settings-preferred-name"
+                />
+                <input data-testid="settings-preferred-name-input" type="hidden" value={preferredName} readOnly />
+
+                {/* Phone field hidden — not currently surfaced as a feature.
+                    State and backend wiring are intentionally left in place
+                    so re-enabling is a one-uncomment change.
+                <TextInput
+                  label="Phone"
+                  value={phone}
+                  onChange={setPhone}
+                  placeholder="+1 555 000 0000"
+                  id="settings-phone"
+                />
+                <input data-testid="settings-phone-input" type="hidden" value={phone} readOnly />
+                */}
+              </div>
+            )}
+
+            {updateProfileMutation.error && (
+              <p data-testid="settings-save-error" className="text-xs font-mono" style={{ color: '#ff6b6b' }}>
+                {updateProfileMutation.error instanceof Error
+                  ? updateProfileMutation.error.message
+                  : "Failed to save"}
+              </p>
+            )}
+
+            {saveSuccess && (
+              <p data-testid="settings-save-success" className="text-xs font-mono" style={{ color: 'var(--c-accent-dim)' }}>
+                Saved.
+              </p>
+            )}
+
+            <Button
+              data-testid="settings-save-button"
+              onClick={handleSave}
+              disabled={updateProfileMutation.isPending}
+              isLoading={updateProfileMutation.isPending}
+              loadingText="Saving…"
+            >
+              Save Changes
+            </Button>
+          </section>
+
+          {/* Email is its own section so the user doesn't mistake it for
+              something the "Save Changes" button covers — email only mutates
+              via the OTP-verified flow below. */}
+          <section className="flex flex-col gap-4 mb-12">
+            <h2 className="text-xs font-mono font-medium uppercase tracking-widest pb-1 border-b" style={{ color: 'var(--c-text-dim)', borderColor: 'var(--c-border)' }}>
+              Email
+            </h2>
+
+            {isLoading ? (
+              <span className="text-xs font-mono" style={{ color: 'var(--c-text-muted)' }}>
+                Loading…
+              </span>
+            ) : emailChangeStep === "idle" ? (
+              <div className="flex flex-col gap-1.5">
+                <TextInput
+                  label="Email"
+                  value={email}
+                  onChange={() => { /* read-only — change via the button below */ }}
+                  type="email"
+                  placeholder="you@example.com"
+                  id="settings-email"
+                  disabled
+                />
+                <input data-testid="settings-email-input" type="hidden" value={email} readOnly />
+                <Button
+                  data-testid="settings-email-change-button"
+                  variant="secondary"
+                  size="sm"
+                  onClick={() => {
+                    setPendingNewEmail("");
+                    setEmailOtpCode("");
+                    setEmailChangeError(null);
+                    setEmailChangeStep("request");
+                  }}
+                  className="self-start mt-3"
+                >
+                  Change Email
+                </Button>
+              </div>
+            ) : emailChangeStep === "request" ? (
+              <div className="flex flex-col gap-2">
+                <TextInput
+                  label="New email"
+                  value={pendingNewEmail}
+                  onChange={setPendingNewEmail}
+                  type="email"
+                  placeholder="you@example.com"
+                  id="settings-email-new"
+                  disabled={emailChangePending}
+                />
+                <input data-testid="settings-email-new-input" type="hidden" value={pendingNewEmail} readOnly />
+                {emailChangeError && (
+                  <p data-testid="settings-email-change-error" className="text-xs font-mono" style={{ color: "#ff6b6b" }}>
+                    {emailChangeError}
+                  </p>
+                )}
+                <div className="flex gap-2">
+                  <Button
+                    data-testid="settings-email-send-code"
+                    size="sm"
+                    onClick={handleSendEmailChangeOtp}
+                    isLoading={emailChangePending}
+                    loadingText="Sending…"
+                  >
+                    Send Code
+                  </Button>
+                  <Button
+                    data-testid="settings-email-cancel"
+                    variant="ghost"
+                    size="sm"
+                    onClick={cancelEmailChange}
+                    disabled={emailChangePending}
+                  >
+                    Cancel
+                  </Button>
+                </div>
+              </div>
+            ) : (
+              <div className="flex flex-col gap-2">
+                <p className="text-xs font-mono" style={{ color: "var(--c-text-muted)" }}>
+                  Code sent to <span style={{ color: "var(--c-text)" }}>{pendingNewEmail}</span>.
+                </p>
+                <TextInput
+                  label="Verification code"
+                  value={emailOtpCode}
+                  onChange={setEmailOtpCode}
+                  placeholder="000000"
+                  id="settings-email-otp"
+                  disabled={emailChangePending}
+                />
+                <input data-testid="settings-email-otp-input" type="hidden" value={emailOtpCode} readOnly />
+                {emailChangeError && (
+                  <p data-testid="settings-email-change-error" className="text-xs font-mono" style={{ color: "#ff6b6b" }}>
+                    {emailChangeError}
+                  </p>
+                )}
+                <div className="flex gap-2">
+                  <Button
+                    data-testid="settings-email-verify"
+                    size="sm"
+                    onClick={handleVerifyEmailChange}
+                    isLoading={emailChangePending}
+                    loadingText="Verifying…"
+                  >
+                    Verify
+                  </Button>
+                  <Button
+                    data-testid="settings-email-back"
+                    variant="ghost"
+                    size="sm"
+                    onClick={() => {
+                      setEmailOtpCode("");
+                      setEmailChangeError(null);
+                      setEmailChangeStep("request");
+                    }}
+                    disabled={emailChangePending}
+                  >
+                    Back
+                  </Button>
+                </div>
+              </div>
+            )}
+          </section>
+
+          {/* Avatar */}
+          <section className="flex flex-col gap-4 mb-12">
+            <h2 className="text-xs font-mono font-medium uppercase tracking-widest pb-1 border-b" style={{ color: 'var(--c-text-dim)', borderColor: 'var(--c-border)' }}>
+              Avatar
+            </h2>
+
+            <div className="flex items-center gap-4">
+              <div
+                data-testid="avatar-preview-container"
+                className="w-14 h-14 overflow-hidden flex items-center justify-center flex-shrink-0 cursor-pointer rounded-panel"
+                style={{ border: '2px solid var(--c-border)', background: 'var(--c-surface-high)' }}
+                onClick={() => fileInputRef.current?.click()}
+                title="Click to choose image"
+              >
+                {preview ? (
+                  <img data-testid="avatar-new-preview" src={preview} alt="Preview" className="w-full h-full object-cover" />
+                ) : currentAvatarUrl ? (
+                  <img
+                    data-testid="avatar-current"
+                    src={currentAvatarUrl}
+                    alt="Avatar"
+                    className="w-full h-full object-cover"
+                    onError={() => setCurrentAvatarUrl(null)}
+                  />
+                ) : (
+                  <User data-testid="avatar-placeholder" size={22} aria-hidden="true" style={{ color: 'var(--c-text-muted)' }} />
+                )}
+              </div>
+
+              <div className="flex flex-col gap-2">
+                <label
+                  htmlFor="settings-avatar-input"
+                  className="inline-flex items-center gap-1.5 text-xs font-mono cursor-pointer transition-colors"
+                  style={{ color: 'var(--c-accent)' }}
+                >
+                  <Upload size={14} aria-hidden="true" />
+                  Choose image
+                </label>
+                <input
+                  key={fileInputKey}
+                  ref={fileInputRef}
+                  id="settings-avatar-input"
+                  data-testid="settings-avatar-input"
+                  type="file"
+                  accept="image/*"
+                  onChange={handleFileChange}
+                  disabled={updateAvatarMutation.isPending}
+                  aria-label="Select avatar image"
+                  className="sr-only"
+                />
+                <p className="text-xs font-mono" style={{ color: 'var(--c-text-muted)' }}>
+                  PNG, JPG, GIF — max 5MB
+                </p>
+              </div>
+            </div>
+
+            {uploadError && (
+              <p data-testid="avatar-upload-error" className="text-xs font-mono" style={{ color: '#ff6b6b' }}>
+                {uploadError}
+              </p>
+            )}
+
+            {saveSuccess && !selectedFile && (
+              <p data-testid="avatar-upload-success" className="text-xs font-mono" style={{ color: 'var(--c-accent-dim)' }}>
+                Avatar updated.
+              </p>
+            )}
+
+            {selectedFile && (
+              <Button
+                data-testid="upload-avatar-button"
+                onClick={handleAvatarUpload}
+                disabled={updateAvatarMutation.isPending}
+                isLoading={updateAvatarMutation.isPending}
+                loadingText="Uploading…"
+              >
+                Upload Avatar
+              </Button>
+            )}
+          </section>
+
+          {/* Software Updates */}
+          <section className="flex flex-col gap-4 mb-12">
+            <h2 className="text-xs font-mono font-medium uppercase tracking-widest pb-1 border-b" style={{ color: 'var(--c-text-dim)', borderColor: 'var(--c-border)' }}>
+              Software Updates
+            </h2>
+
+            <div className="flex flex-col gap-2">
+              <p className="text-xs font-mono" style={{ color: 'var(--c-text-muted)' }}>
+                Current version: <span style={{ color: 'var(--c-text)' }}>{appVersion || "Loading..."}</span>
+              </p>
+
+              {updateStatus === "available" && (
+                <p className="text-xs font-mono" style={{ color: 'var(--c-accent)' }}>
+                  Update available: {updateVersion}
+                </p>
+              )}
+
+              {updateStatus === "none" && (
+                <p className="text-xs font-mono" style={{ color: 'var(--c-accent-dim)' }}>
+                  You're up to date!
+                </p>
+              )}
+
+              {updateStatus === "error" && (
+                <p className="text-xs font-mono" style={{ color: '#ff6b6b' }}>
+                  {updateError}
+                </p>
+              )}
+            </div>
+
+            <div className="flex gap-2">
+              <Button
+                onClick={handleCheckForUpdates}
+                disabled={updateStatus === "checking"}
+                isLoading={updateStatus === "checking"}
+                loadingText="Checking…"
+              >
+                Check for updates
+              </Button>
+
+              {updateStatus === "available" && (
+                <Button
+                  onClick={handleInstallUpdate}
+                  disabled={isInstalling}
+                  isLoading={isInstalling}
+                  loadingText="Installing…"
+                  variant="primary"
+                >
+                  Install update
+                </Button>
+              )}
+            </div>
+          </section>
+
+        </div>
+      </div>
+    </div>
     </PageShell>
   );
 };
